Methods of Charging and Functions of Charge
Included snippits of charged drink as a method and a cost message for the final display. Created a Loop so the drink order repeats and a function for charges. Will fix the logic: I have the general idea how the code should go just need to re-organize whether I want to make the calculation portion a method in the class or a function in the manager.